fix rename folders V1.13.1
