+++ pangeas-v1/src/components/map/MapLeaflet.vue
@@ -7,8 +7,8 @@ import L from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import 'leaflet-routing-machine';
 import { nextTick } from 'vue';
-import PlaceModal from '../Map/PlaceModal.vue';
-import ValidateVisitPopup from '../Modal/ValidateVisitPopup.vue';
+import PlaceModal from './/PlaceModal.vue';
+import ValidateVisitPopup from '../modal/ValidateVisitPopup.vue';
 import haversine from 'haversine-distance';
 import store from '../../store';
 import { createApp } from 'vue';
