+++ pangeas-v1/src/components/map/PlaceDetail.vue
