+++ pangeas-v1/src/components/modal/DeleteAccountModal.vue
