+++ pangeas-v1/src/components/modal/EmailUpdateModal.vue
