+++ pangeas-v1/src/components/modal/Registration.vue
